Move legacy slurmconf cache file conversion into a script
